Add blueprints and embedded records documentation
This moves the base app adapter and serializer to a special module
called "drf" and then extends that module in the as "application".  This
allows us to do last minute tweaks behind the scenes at the app-level,
while still allowing the user to extend the adapter/serializer to add
their own customizations.

Additionally, this adds documentation for working with embedded records
which requires the user the use the new blueprint feature.
