+++ app/adapters/application.js
@@ -1,16 +1,3 @@
-import DRFAdapter from 'ember-django-adapter/adapters/drf';
-import ENV from '../config/environment';
+import DRFAdapter from './drf';
 
-export default DRFAdapter.extend({
-  host: function() {
-    return ENV.APP.API_HOST;
-  }.property(),
-
-  namespace: function() {
-    return ENV.APP.API_NAMESPACE;
-  }.property(),
-
-  add_trailing_slashes: function() {
-    return ENV.APP.API_ADD_TRAILING_SLASHES;
-  }.property()
-});
+export default DRFAdapter;
