+++ app/serializers/drf.js
@@ -1,0 +1,3 @@
+import DRFSerializer from 'ember-django-adapter/serializers/drf';
+
+export default DRFSerializer;
