+++ blueprints/.jshintrc
@@ -1,0 +1,5 @@
+{
+  "predef": [
+    "console"
+  ]
+}
