+++ blueprints/drf-adapter/files/app/adapters/__name__.js
@@ -1,0 +1,4 @@
+import DRFAdapter from './drf';
+
+export default DRFAdapter.extend({
+});
