+++ blueprints/drf-serializer/files/app/serializers/__name__.js
@@ -1,0 +1,4 @@
+import DRFSerializer from './drf';
+
+export default DRFSerializer.extend({
+});
