+++ blueprints/drf-serializer/index.js
@@ -1,0 +1,14 @@
+module.exports = {
+  description: ''
+
+  // locals: function(options) {
+  //   // Return custom template variables here.
+  //   return {
+  //     foo: options.entity.options.foo
+  //   };
+  // }
+
+  // afterInstall: function(options) {
+  //   // Perform extra work here.
+  // }
+};
